move some things around relating to how the admin part of srv is organised
