+++ martin/src/srv/mod.rs
@@ -2,17 +2,17 @@
 mod fonts;
 
 mod server;
-pub use server::{Catalog, RESERVED_KEYWORDS, new_server, router};
+pub use server::{new_server, RESERVED_KEYWORDS, router};
+
+mod admin;
+pub use admin::Catalog;
 
 #[cfg(feature = "_tiles")]
 mod tiles;
 #[cfg(feature = "_tiles")]
-pub use tiles::{DynTileSource, TileRequest};
-
+pub use tiles::content::{DynTileSource};
 #[cfg(feature = "_tiles")]
-mod tiles_info;
-#[cfg(feature = "_tiles")]
-pub use tiles_info::merge_tilejson;
+pub use tiles::metadata::merge_tilejson;
 
 #[cfg(feature = "sprites")]
 mod sprites;
