+++ martin/src/srv/tiles/metadata.rs
@@ -191,8 +191,9 @@ pub mod tests {
 
     use tilejson::{Bounds, VectorLayer};
 
+    use crate::srv::tiles::tests::TestSource;
+
     use super::*;
-    use crate::srv::server::tests::TestSource;
 
     #[test]
     fn test_merge_tilejson() {
